+++ .prettierrc
@@ -1,4 +1,7 @@
 {
   "singleQuote": true,
-  "trailingComma": "all"
+  "trailingComma": "all",
+  "printWidth": 120,
+  "semi": false,
+  "prettier.bracketSpacing": false
 }
