+++ src/Presentation/Sample/app.controller.spec.ts
@@ -1,0 +1,22 @@
+import { Test, TestingModule } from '@nestjs/testing'
+import { AppController } from './app.controller'
+import { AppService } from '../../Application/Sample/app.service'
+
+describe('AppController', () => {
+  let appController: AppController
+
+  beforeEach(async () => {
+    const app: TestingModule = await Test.createTestingModule({
+      controllers: [AppController],
+      providers: [AppService],
+    }).compile()
+
+    appController = app.get<AppController>(AppController)
+  })
+
+  describe('root', () => {
+    it('should return "Hello World!"', () => {
+      expect(appController.getHello()).toBe('Hello World!')
+    })
+  })
+})
