+++ src/Presentation/Sample/app.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common'
+import { AppService } from '../../Application/Sample/app.service'
+
+@Controller()
+export class AppController {
+  constructor(private readonly appService: AppService) {}
+
+  @Get()
+  getHello(): string {
+    return this.appService.getHello()
+  }
+}
